refactor(frontend): rename Ethereum send review component
